fixing ipad view
