+++ SongLib/Data/Sample/SampleSongs.swift
@@ -1,5 +1,5 @@
 //
-//  SampleBooks.swift
+//  SampleSongs.swift
 //  SongLib
 //
 //  Created by Siro Daves on 15/08/2025.
