+++ SongLib/Data/Sample/SampleVerses.swift
@@ -1,0 +1,16 @@
+//
+//  SampleVerses.swift
+//  SongLib
+//
+//  Created by Siro Daves on 15/08/2025.
+//
+
+extension String {
+    static let sampleVerses:  [String] = [
+        "When the great plants of our cities\nhave turned out their last finished work,\nWhen our merchants have sold\ntheir last yard of goods and dismissed\nthe last tired clerk;\nWhen our banks have raked in their\nlast dollar\nAnd paid the last dividends,\nWhen the Judge of the earth says,\n'Closed for the night,'\nAnd asks for a balance..\nWhat Then?",
+        "CHORUS\nWhat then, What then?\nWhen the great Book is opened, what then?\nWhen the ones that’s rejecting\nthis Message tonight,\nWill be asked to give a reason -\nWhat Then?",
+        "When the choir has sung its last song\nand the preacher has said his last prayer,\nWhen the people have heard their last\nsermon\nAnd the sound has died out in the air,\nWhen the Bible lies closed on the altar\nAnd the pews are all empty of folks,\nAnd each one stands facing his record\nAnd the great Book is opened..\nWhat Then?",
+        "When the actors have played their last drama\nAnd the mimic has made his last fun,\nWhen the film has flashed its last film\nAnd the billboard displayed its last run,\nWhen the crowds seeking pleasure have vanished\nAnd all the liquor stores and beer joints have been closed,\nAnd the salesman sold his last drink\nAnd gone out in the darkness,\nWhen the trumpet of ages is sounded\nAnd we stand before Him..\nWhat Then?",
+        "When the bugle call sinks into silence\nand the long marching columns stand still,\nAnd the captain repeats his last orders,\nAnd the flag has been hauled from the mast,\nAnd the wounded afield checked in\nand a world that rejects its Saviour\nIs asked for a reason..\nWhat Then?"
+    ]
+}
