--- SongLib/Presentation/Views/Screens/Selection/Step1View.swift
@@ -1,148 +1,0 @@
-//
-//  Step1View.swift
-//  SongLib
-//
-//  Created by Siro Daves on 30/04/2025.
-//
-
-import SwiftUI
-
-struct Step1View: View {
-    @StateObject private var viewModel: SelectionViewModel = DiContainer.shared.resolve(SelectionViewModel.self)
-    @State private var showNoSelectionAlert = false
-    @State private var showConfirmationAlert = false
-    @State private var navigateToNextScreen = false
-
-    var body: some View {
-        Group {
-            if navigateToNextScreen {
-                AnyView(Step2View())
-            } else {
-                AnyView(mainContent)
-            }
-        }
-    }
-    
-    private var mainContent: some View {
-        VStack {
-            Text("Select Songbooks")
-                .font(.title2)
-                .foregroundColor(.onPrimaryContainer)
-            stateContent.background(.surface)
-        }
-        .background(.primaryContainer)
-        .alert(isPresented: $showNoSelectionAlert) {
-            noSelectionAlert
-        }
-        .confirmationDialog(
-            "If you are done selecting please proceed ahead. We can always bring you back here to reselect afresh.",
-            isPresented: $showConfirmationAlert,
-            titleVisibility: .visible
-        ) {
-            confirmationDialogActions
-        }
-        .task({viewModel.fetchBooks()})
-        .onChange(of: viewModel.uiState, perform: handleStateChange)
-    }
-    
-    @ViewBuilder
-    private var stateContent: some View {
-        switch viewModel.uiState {
-            case .loading(let msg):
-                LoadingState(
-                    title: msg ?? "Loading books ...",
-                    fileName: "loading-hand"
-                )
-                
-            case .saving(let msg):
-            LoadingState(
-                title: msg ?? "Loading books ...",
-                fileName: "cloud-download"
-            )
-                
-            case .saved:
-                LoadingView()
-                
-            case .error(let msg):
-                ErrorView(message: msg) {
-                    Task { viewModel.fetchBooks() }
-                }
-                
-            default:
-                BookSelectionView(
-                    viewModel: viewModel,
-                    showNoSelectionAlert: $showNoSelectionAlert,
-                    showConfirmationAlert: $showConfirmationAlert
-                )
-        }
-    }
-    
-    private var noSelectionAlert: Alert {
-        Alert(
-            title: Text("Oops! No selection found"),
-            message: Text("Please select at least 1 song book to proceed to the next step."),
-            dismissButton: .default(Text("OK"))
-        )
-    }
-    
-    private var confirmationDialogActions: some View {
-        Group {
-            Button("Proceed") {
-                viewModel.saveBooks()
-            }
-            Button("Cancel", role: .cancel) {}
-        }
-    }
-    
-    private func handleStateChange(_ state: UiState) {
-        navigateToNextScreen = .saved == state
-    }
-}
-
-struct BookSelectionView: View {
-    @ObservedObject var viewModel: SelectionViewModel
-    @Binding var showNoSelectionAlert: Bool
-    @Binding var showConfirmationAlert: Bool
-
-    var body: some View {
-        VStack {
-            ScrollView {
-                LazyVStack(spacing: 12) {
-                    ForEach(viewModel.books.indices, id: \.self) { index in
-                        let selectable = viewModel.books[index]
-                        BookItem(
-                            book: selectable.data,
-                            isSelected: selectable.isSelected
-                        ) {
-                            viewModel.toggleSelection(for: selectable.data)
-                        }
-                    }
-                }
-                .padding()
-            }
-
-            Button(action: {
-                if viewModel.selectedBooks().isEmpty {
-                    showNoSelectionAlert = true
-                } else {
-                    showConfirmationAlert = true
-                }
-            }) {
-                HStack(spacing: 5) {
-                    Image(systemName: "checkmark")
-                    Text("Proceed")
-                }
-                .frame(width: 150)
-                .foregroundColor(.onPrimaryContainer)
-                .padding()
-                .background(.primaryContainer)
-                .cornerRadius(10)
-            }
-            .padding(.bottom)
-        }
-    }
-}
-
-#Preview {
-    Step1View()
-}
