+++ SongLib/Presentation/Views/Screens/Selection/Step2/Step2View.swift
@@ -28,7 +28,7 @@ struct Step2View: View {
         VStack {
             stateContent
         }
-        .background(.primaryContainer)
+//        .background(.primaryContainer)
         .onAppear {
             viewModel.initializeStep2()
         }
